simulacrum::store::in_mem_store

Struct InMemoryStore

pub struct InMemoryStore { /* private fields */ }

Implementations

impl InMemoryStore

pub fn new(genesis: &Genesis) -> Self

pub fn get_checkpoint_by_sequence_number( &self, sequence_number: CheckpointSequenceNumber, ) -> Option<&VerifiedCheckpoint>

pub fn get_checkpoint_by_digest( &self, digest: &CheckpointDigest, ) -> Option<&VerifiedCheckpoint>

pub fn get_highest_checkpoint(&self) -> Option<&VerifiedCheckpoint>

pub fn get_checkpoint_contents( &self, digest: &CheckpointContentsDigest, ) -> Option<&CheckpointContents>

pub fn get_committee_by_epoch(&self, epoch: EpochId) -> Option<&Committee>

pub fn get_transaction( &self, digest: &TransactionDigest, ) -> Option<&VerifiedTransaction>

pub fn get_transaction_effects( &self, digest: &TransactionDigest, ) -> Option<&TransactionEffects>

pub fn get_transaction_events( &self, digest: &TransactionDigest, ) -> Option<&TransactionEvents>

pub fn get_object(&self, id: &ObjectID) -> Option<&Object>

pub fn get_object_at_version( &self, id: &ObjectID, version: SequenceNumber, ) -> Option<&Object>

pub fn get_system_state(&self) -> IotaSystemState

pub fn get_clock(&self) -> Clock

pub fn get_last_checkpoint_of_epoch( &self, epoch: EpochId, ) -> Option<CheckpointSequenceNumber>

pub fn get_system_state_by_epoch( &self, epoch: EpochId, ) -> Option<&IotaSystemState>

Get the system state for a specific epoch. Returns None if the system state for that epoch was not stored.

pub fn store_system_state_for_epoch( &mut self, epoch: EpochId, system_state: IotaSystemState, )

Store the system state for a specific epoch. This should be called when an epoch ends to preserve the final system state.

pub fn owned_objects(&self, owner: IotaAddress) -> impl Iterator<Item = &Object>

pub fn update_last_checkpoint_by_epoch( &mut self, epoch: EpochId, last_checkpoint: CheckpointSequenceNumber, )

impl InMemoryStore

pub fn insert_checkpoint(&mut self, checkpoint: VerifiedCheckpoint)

pub fn insert_checkpoint_contents(&mut self, contents: CheckpointContents)

pub fn insert_committee(&mut self, committee: Committee)

pub fn insert_executed_transaction( &mut self, transaction: VerifiedTransaction, effects: TransactionEffects, events: TransactionEvents, written_objects: BTreeMap<ObjectID, Object>, )

pub fn insert_transaction(&mut self, transaction: VerifiedTransaction)

pub fn insert_transaction_effects(&mut self, effects: TransactionEffects)

pub fn insert_events( &mut self, tx_digest: &TransactionDigest, events: TransactionEvents, )

pub fn update_objects( &mut self, written_objects: BTreeMap<ObjectID, Object>, deleted_objects: Vec<ObjectRef>, )