Crate simulacrum

A Simulacrum of Sui.

The word simulacrum is latin for “likeness, semblance”, it is also a spell in D&D which creates a copy of a creature which then follows the player’s commands and wishes. As such this crate provides the Simulacrum type which is a implementation or instantiation of a sui blockchain, one which doesn’t do anything unless acted upon.

Re-exports

pub use self::store::SimulatorStore;

pub use self::store::in_mem_store::InMemoryStore;

Modules

store

Structs

AdvanceEpochConfig

Configuration for advancing epochs in the Simulacrum.

CommitteeWithKeys

Simulacrum

A Simulacrum of Sui.